refactor(frontend): remove barrel imports from store
- Replace @/lib/store with direct imports from specific files
- useAuthStore from @/lib/store/auth-store
- useUIStore from @/lib/store/ui-store
- useServersStore from @/lib/store/servers-store
- Improves HMR speed (avoids re-processing all exports)
- Affected: 4 files
